Create manage-list-model.js
In this code, we import and use the following packages:
express: A web framework for Node.js
body-parser: A middleware for parsing incoming request bodies
We create a new Express app and add the following middleware:
body-parser.json: Parses incoming JSON payloads
We create a new array to store our list items and define the following routes:
GET /items: Returns the current list of items as JSON
POST /items: Adds a new item to the list and returns the new item as JSON
PUT /items/:id: Updates an existing item with the specified ID and returns the updated item as JSON
DELETE /items/:id: Deletes an item with the specified ID and returns a 204 No Content status
Note that in this example, we are not using any database to persist the list items, so they will be lost when the server is restarted. If you need to persist data between server sessions, you can use a database or a file system.