Create ajax-request-model.py
o make an AJAX request in Node.js, you can use the built-in http or https modules or third-party libraries such as axios, node-fetch, or request.
In this example, we are making a GET request to http://example.com/api/data and logging the response to the console. We are using the http.get() method to initiate the request, which takes two arguments: the URL to request and a callback function to handle the response.

Inside the callback function, we create an empty string data to store the response body. We add to this string each time we receive a new chunk of data using the data event of the res object. When the response is finished, we parse the data string as JSON and log it to the console.

You can also use the http.request() method to make other types of requests (e.g. POST, PUT, DELETE) or to set request headers and options. The http module provides low-level access to the HTTP protocol and can be used to build more complex request and response handlers.

Note that this example is for making requests from Node.js on the server-side. If you want to make AJAX requests from a web browser, you would typically use the XMLHttpRequest object or the fetch() API, which are built into modern web browsers.