Create ajax-post-model.js
In this example, we use the axios library to send an AJAX POST request to the URL https://jsonplaceholder.typicode.com/posts with a JSON payload containing a title, body, and userId. We then log the response data to the console. If an error occurs, we log the error to the console.

Note that you can replace the URL and JSON payload with your own values depending on your specific use case.